refactor(mm-next): rename components under login page
